Revert "Animacion"
