--- Assets/Animaciones.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ba52d7586e03e3443915afd7a03a5d87
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
