--- Assets/Animaciones/Animation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ce667af922461084682be52444f709d2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
