--- Assets/Animaciones/Controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 20c32d0e41991c942acfbefaebb7ccd0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
