--- Assets/Animaciones/Controller/AnimadorBebe.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 20148d6db945a0a4b9e8193cf8aeda2b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
