--- Assets/Prefab/Modelos.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1652b27bdab78ca4ba21006fe08b77d2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
