--- Assets/Prefab/Modelos/bebe.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 960220a56eb6ad948900cffdcfa92e48
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
