--- Assets/Scripts/AnimacionTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e9a36b7e9c560cc418fdfe0f6287a5b5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
