--- Assets/Scripts/Estado.cs
@@ -1,5 +1,0 @@
-﻿
-    public enum Estado
-    {
-    MOVE,PUSH
-    }
