--- Assets/Scripts/Estado.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0780a1d1615c3234da74c9488877325d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
