--- Assets/Scripts/Interactuable.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 169c2e3fbb4853845adfce89b813127f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
