--- Assets/Scripts/Interactuable/Interactuable.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface Interactuable 
-{
-    void OnInteraction();
-    bool subible();
-    bool interactuable();
-}
