--- Assets/Scripts/Interactuable/Interactuable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d296c9b614fd81e41a66bc735cae566b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
