--- Assets/Scripts/Interactuable/empujable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7a683b45138ec434c89b2249fb062322
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
