--- Assets/Scripts/Player/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2b4152d812575f1438e0d4c73308b6c6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
