--- Assets/Scripts/Player/move.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d46738add6e53674ebb05a99c56786b0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
