--- Assets/Scripts/UI/buttonFunction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 028b5c890d66d7741abba689af165064
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
