--- Assets/Scripts/UI/menupausa.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7cf57f8c998ed7b42a405dff1cde2513
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
