+++ Assets/Scripts/moveCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 263941b7b140d5a4b96927e0e7b827b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
