+++ Assets/Scripts/takeObject.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-
 public class takeObject : MonoBehaviour
 {
 
@@ -11,7 +10,10 @@ public class takeObject : MonoBehaviour
 
 
 
-    
+    Ray ray;
+    RaycastHit hit;
+    GameObject obj;
+    public float distance;
 
     // Start is called before the first frame update
     void Start()
@@ -19,9 +21,25 @@ public class takeObject : MonoBehaviour
         State = Mode.PLAY;
     }
 
-    void FixedUpdate()
+    // Update is called once per frame
+    void Update()
     {
-        
+        if (State == Mode.PLAY)
+        {
+            ray = Camera.main.ScreenPointToRay(Input.mousePosition);
+            if (Physics.Raycast(ray, out hit,distance))
+            {
+                obj = hit.collider.gameObject;
+
+                if (obj.tag == "KeyObject")
+                {
+                    if(Input.GetButton("interactuar"))
+                    {
+                        Destroy(obj);
+                    }
+                }
+            }
+        }
     }
